+++ test/workout_files_test.dart
@@ -1,6 +1,5 @@
 import 'dart:io';
 
-import 'package:csv/csv.dart';
 import 'package:fit_tool/fit_tool.dart';
 import 'package:fit_tool/src/fit_file.dart';
 import 'package:test/test.dart';
@@ -48,6 +47,12 @@ void main() {
           }
         }
       }
+    });
+
+    test('Decode workout file with multiple dev data indexes', () async {
+      final file = File('./test/data/stagesLink_25622166.fit');
+      final bytes = await file.readAsBytes();
+      final fitFile = FitFile.fromBytes(bytes);
     });
   });
 }